+++ npm-packages/meteor-vite/src/meteor/IPC/methods/vite-server.ts
@@ -1,7 +1,7 @@
 import FS from 'fs/promises';
 import Path from 'path';
 import { createServer, resolveConfig, ViteDevServer } from 'vite';
-import Logger from '../../../Logger';
+import Logger from '../../../utilities/Logger';
 import MeteorEvents, { MeteorIPCMessage } from '../MeteorEvents';
 import { MeteorViteConfig } from '../../../vite/MeteorViteConfig';
 import { MeteorStubs } from '../../../vite';
