+++ npm-packages/meteor-vite/src/meteor/package/components/PackageExport.ts
@@ -1,6 +1,6 @@
 import pc from 'picocolors';
 import PackageJSON from '../../../../package.json';
-import Logger from '../../../Logger';
+import Logger from '../../../utilities/Logger';
 import { PACKAGE_SCOPE_KEY, TEMPLATE_GLOBAL_KEY } from '../StubTemplate';
 import MeteorPackage from './MeteorPackage';
