+++ npm-packages/meteor-vite/src/meteor/parser/Parser.ts
@@ -9,7 +9,7 @@ import {
 } from '@babel/types';
 import FS from 'fs/promises';
 import { inspect } from 'util';
-import Logger from '../../Logger';
+import Logger from '../../utilities/Logger';
 import { MeteorViteError } from '../../vite/error/MeteorViteError';
 import {
     KnownModuleMethodNames,
