+++ npm-packages/meteor-vite/src/utilities/GenericTypes.ts
