+++ npm-packages/meteor-vite/src/vite/error/ErrorHandler.ts
@@ -1,4 +1,4 @@
-import Logger from '../../Logger';
+import Logger from '../../utilities/Logger';
 import ViteLoadRequest, { RefreshNeeded } from '../ViteLoadRequest';
 import { MeteorViteError } from './MeteorViteError';
